Basic types.
These are heavily inspired by Criterion. The main difference is that
unlike Criterion, the `Benchmarkable` type is a monad. This makes it
possible to reduce the function space for `env` significantly, avoiding
exotic functions that return a different set of benchmarks by inspecting
the resource value. Exotic functions are not just a theoretical problem
in Criterion: the framework passes `undefined` to get the list of all
benchmarks. So being strict on the argument in any way makes the whole
thing go boom.