Force Bracket resource
`Bracket` imposes an `NFData` constraint on the resource. The idea is to
ensure that the input is fully evaluated before it is used in the
benchmark. However the value was never forced.

This patch forces the evaluation after the resource initialization, but
before the value is passed to the benchmark.