Remove Env in Series constructor
Prior to this change `Series` took an `Env` argument for running the
benchmark, similar to `Bracket`. The `Env` wrapper makes it impossible
for the user to inspect the parameter when building the `Benchmark`; the
reason for this is that we want to be able to list the benchmarks
"purely", without creating the resource. However in `Series` the
benchmark input does not need to be initialized so we can access it
purely.