Introduce indexed sampling strategies.
This is a first step towards implementing filtering benchmarks, since
the caller to `runBenchmarks` can pass in a function that returns no
strategy for some benchmark identifiers. The new type of
`runBenchmark` is general enough that callers can decide to use
different sampling strategies for different benchmarks. But we
probably won't be using the full power allowed by the type any time
soon.

Furthermore, we make it possible to set a uniform sampling strategy in
the ConfigMonoid passed to `defaultMainWith`. Uniform sampling
strategies are those that sample in the same way all benchmarks.