Don't duplicate report by default
Before these changes the default behavior would be to print both a
pretty-printed and a json report, which doesn't make much sense. This
makes the pretty-printing the default.

It introduces a new datatype `ReportOutput` which records what report
format should be printed. This should also make it easier to add new
report formats.

_Note: this breaks compatibility: `--output` is now `--json`._